examples/strands_math_agent/pyproject.toml
[project]
name = "strands-math-agent-example"
version = "0.1.0"
description = "Example: Strands Math Agent using Bedrock AgentCore RL Toolkit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=1.0.3",
    "bedrock-agentcore-starter-toolkit<=0.2.0",
    "boto3>=1.40.55",
    "python-dotenv>=1.0.0",
    # TODO: replace the above dependencies with agentcore-rl-toolkit>=0.1.0 after PyPI indexing
    "strands-agents[openai]>=1.18.0",
    "strands-agents-tools>=0.2.16",
]

[tool.setuptools]
py-modules = ["basic_app", "rl_app", "reward"]
